Breaking: Move `content-type` util from `hint` to `@hint/utils`
Also update `content-type` to reference trimmed `mime-db`.
