+++ packages/connector-jsdom/src/connector.ts
@@ -31,8 +31,8 @@ import { fork, ChildProcess } from 'child_process';
 
 import { JSDOM, ResourceLoader, VirtualConsole } from 'jsdom';
 
-import { debug as d, dom, HTMLElement, HTMLDocument, HttpHeaders, network } from '@hint/utils';
-import { Engine, Event, FetchEnd, FetchError, IConnector, NetworkData, utils } from 'hint';
+import { contentType, debug as d, dom, HTMLElement, HTMLDocument, HttpHeaders, network } from '@hint/utils';
+import { Engine, Event, FetchEnd, FetchError, IConnector, NetworkData } from 'hint';
 import { Requester } from '@hint/utils-connector-tools';
 
 import CustomResourceLoader from './resource-loader';
@@ -45,7 +45,7 @@ import { beforeParse } from './before-parse';
  */
 
 const { createHTMLDocument, traverse } = dom;
-const { contentType: { getContentTypeData, getType } } = utils;
+const { getContentTypeData, getType } = contentType;
 const { isHTMLDocument } = network;
 const debug: debug.IDebugger = d(__filename);
