+++ packages/hint/package.json
@@ -25,7 +25,6 @@
     "is-svg": "^4.1.0",
     "jsonc-parser": "^2.1.0",
     "lodash": "^4.17.11",
-    "mime-db": "1.35.0",
     "optionator": "^0.8.2",
     "ora": "^3.4.0",
     "semver": "^6.0.0",
@@ -36,7 +35,6 @@
     "@types/async": "^2.4.1",
     "@types/content-type": "^1.1.3",
     "@types/lodash": "^4.14.123",
-    "@types/mime-db": "^1.27.0",
     "@types/node": "^11.13.4",
     "@types/semver": "^6.0.0",
     "@types/sinon": "^7.0.11",
