+++ packages/hint/src/lib/utils/index.ts
@@ -1,5 +1,4 @@
 import * as packagesUtils from './packages';
-import * as contentTypeUtils from './content-type';
 import * as resourceLoaderUtils from './resource-loader';
 import * as schemaValidatorUtils from './schema-validator';
 import * as jsonParserUtils from './json-parser';
@@ -7,7 +6,6 @@ import * as jsonParserUtils from './json-parser';
 export * from './resource-loader';
 
 export const packages = packagesUtils;
-export const contentType = contentTypeUtils;
 export const resourceLoader = resourceLoaderUtils;
 export const schemaValidator = schemaValidatorUtils;
 export const jsonParser = jsonParserUtils;
