--- packages/hint/tests/lib/utils/fixtures/dummy.txt
@@ -1,1 +1,0 @@
-dummy
