--- packages/hint/tests/lib/utils/fixtures/fixture.json
@@ -1,3 +1,0 @@
-{
-    "property1": "value1"
-}
