+++ packages/utils-debugging-protocol-common/src/request-response.ts
@@ -1,13 +1,14 @@
 import { atob } from 'abab';
 import { Crdp } from 'chrome-remote-debug-protocol';
 
-import { debug as d, HTMLElement, HttpHeaders } from '@hint/utils';
-import { Response, utils } from 'hint';
+import { contentType, debug as d, HTMLElement, HttpHeaders } from '@hint/utils';
+import { Response } from 'hint';
+
 import { normalizeHeaders, Requester } from '@hint/utils-connector-tools';
 
 const debug: debug.IDebugger = d(__filename);
 
-const { contentType: { getContentTypeData } } = utils;
+const { getContentTypeData } = contentType;
 
 export enum RequestStatus {
     willBeSent = 'willBeSent',
