+++ packages/utils/package.json
@@ -26,6 +26,7 @@
     "@types/configstore": "^4.0.0",
     "@types/debug": "^4.1.4",
     "@types/file-url": "^2.0.0",
+    "@types/mime-db": "^1.27.0",
     "@types/parse5": "^5.0.0",
     "@types/parse5-htmlparser2-tree-adapter": "^5.0.1",
     "@types/request": "^2.48.1",
@@ -39,7 +40,7 @@
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-markdown": "^1.0.0",
     "eventemitter2": "^5.0.1",
-    "mime-db": "^1.39.0",
+    "mime-db": "1.35.0",
     "npm-run-all": "^4.1.5",
     "nyc": "^13.3.0",
     "proxyquire": "2.0.0",
