+++ packages/utils/tests/fixtures/fixture.ico
